requirements.txt
streamlit==1.41.1
numpy==2.0.2
Pillow==11.1.0
tensorflow==2.18.0
keras==3.7.0